python3-anaconda: fix ks file erro with user groups
When fill the "Group Membership" in "ADVANCED USER CONFIGURATION"
dialog with the provided example text:

wheel, my-team (1245), project-x (29935)

It keeps the spaces between group name(project-x) and group id('(29935)'),
and then write them to kickstart file. When boot anaconda with the kickstart
file, it fails with:

| unrecognzed arguments: (1245),project-x (29935)

Filter out the spaces between group name and group id to avoid such
issue.

Signed-off-by: Kai Kang <kai.kang@windriver.com>