python3-anaconda: reset default autopart type
It introduced a new variable default_scheme to replace the old one
DEFAULT_AUTOPART_TYPE. Both of the default values are 'LVM'. But it is
overridden by 'BTRFS' in fedora.conf which is used as default profile
for OE. Because it fails to create home directory for new users, reset
to use 'LVM' to fix the issue and for backward compatile as well.

Signed-off-by: Kai Kang <kai.kang@windriver.com>