python3-anaconda: rebase patch 0078 and add fedora.css
It splits setting of anaconda-gtk.css to linux distros' logo package,
such as fedora-logos. But it is not suitable for meta-anaconda. Copy the
'fedora.css' from fedora-logos package, and adjust the y-axis position
of logo image that replace fixed '20px' with '10%' for better looking.

Rebase 0078-anaconda.py-use-Fedora-as-default-product.patch as well.

Signed-off-by: Kai Kang <kai.kang@windriver.com>